[Feat] delete Exception 기능 개발

+++ checking/src/main/java/sheetplus/checkings/domain/service/ContestCRUDService.java
@@ -57,8 +57,11 @@ public class ContestCRUDService {
 
     @Transactional
     public void deleteContest(Long id) {
-        contestRepository.deleteById(id);
-
+        if(contestRepository.existsById(id)){
+            contestRepository.deleteById(id);
+        }else{
+            throw new ApiException(CONTEST_NOT_FOUND);
+        }
     }
